concatenate patchers
