+++ .ci/conformance/download_tests.sh
@@ -1,14 +1,63 @@
 #!/bin/bash
+while [[ $# -gt 0 ]]; do
+    key="$1"
 
-wget -O test_adaboost.py https://raw.githubusercontent.com/scikit-learn/scikit-learn/0.21.3/sklearn/ensemble/tests/test_weight_boosting.py
-wget -O test_dbscan.py https://raw.githubusercontent.com/scikit-learn/scikit-learn/0.21.3/sklearn/cluster/tests/test_dbscan.py
-wget -O test_elastic_net.py https://raw.githubusercontent.com/scikit-learn/scikit-learn/0.21.3/sklearn/linear_model/tests/test_coordinate_descent.py
-wget -O test_forest.py https://raw.githubusercontent.com/scikit-learn/scikit-learn/0.21.3/sklearn/ensemble/tests/test_forest.py
-wget -O test_kmeans.py https://raw.githubusercontent.com/scikit-learn/scikit-learn/0.21.3/sklearn/cluster/tests/test_k_means.py
-wget -O test_knn.py https://raw.githubusercontent.com/scikit-learn/scikit-learn/0.21.3/sklearn/neighbors/tests/test_neighbors.py
-wget -O test_lin_reg.py https://raw.githubusercontent.com/scikit-learn/scikit-learn/0.21.3/sklearn/linear_model/tests/test_base.py
-wget -O test_log_reg.py https://raw.githubusercontent.com/scikit-learn/scikit-learn/0.21.3/sklearn/linear_model/tests/test_logistic.py
-wget -O test_pca.py https://raw.githubusercontent.com/scikit-learn/scikit-learn/0.21.3/sklearn/decomposition/tests/test_pca.py
-wget -O test_ridge_reg.py https://raw.githubusercontent.com/scikit-learn/scikit-learn/0.21.3/sklearn/linear_model/tests/test_ridge.py
-wget -O test_svm.py https://raw.githubusercontent.com/scikit-learn/scikit-learn/0.21.3/sklearn/svm/tests/test_svm.py
-wget -O test_tree.py https://raw.githubusercontent.com/scikit-learn/scikit-learn/0.21.3/sklearn/tree/tests/test_tree.py
+    case $key in
+        --alg-name)
+        ALG_NAME="$2"
+        ;;
+        --scikit-version)
+        SKLEARN_VER="$2"
+        ;;
+        *)
+        echo "Unknown option: $1"
+        exit 1
+        ;;
+    esac
+    shift
+    shift
+done
+SKLEARN_URL_ROOT="https://raw.githubusercontent.com/scikit-learn/scikit-learn/${SKLEARN_VER}/sklearn/"
+
+case ${ALG_NAME} in
+    "adaboost")
+        wget -O test_adaboost.py ${SKLEARN_URL_ROOT}ensemble/tests/test_weight_boosting.py
+    ;;
+    "dbscan")
+        wget -O test_dbscan.py ${SKLEARN_URL_ROOT}cluster/tests/test_dbscan.py
+    ;;
+    "elastic_net")
+        wget -O test_elastic_net.py ${SKLEARN_URL_ROOT}linear_model/tests/test_coordinate_descent.py
+    ;;
+    "forest")
+        wget -O test_forest.py ${SKLEARN_URL_ROOT}ensemble/tests/test_forest.py
+    ;;
+    "kmeans")
+        wget -O test_kmeans.py ${SKLEARN_URL_ROOT}cluster/tests/test_k_means.py
+    ;;
+    "knn")
+        wget -O test_knn.py ${SKLEARN_URL_ROOT}neighbors/tests/test_neighbors.py
+    ;;
+    "lin_reg")
+        wget -O test_lin_reg.py ${SKLEARN_URL_ROOT}linear_model/tests/test_base.py
+    ;;
+    "log_reg")
+        wget -O test_log_reg.py ${SKLEARN_URL_ROOT}linear_model/tests/test_logistic.py
+    ;;
+    "pca")
+        wget -O test_pca.py ${SKLEARN_URL_ROOT}decomposition/tests/test_pca.py
+    ;;
+    "ridge_reg")
+        wget -O test_ridge_reg.py ${SKLEARN_URL_ROOT}linear_model/tests/test_ridge.py
+    ;;
+    "svm")
+        wget -O test_svm.py ${SKLEARN_URL_ROOT}svm/tests/test_svm.py
+    ;;
+    "tree")
+        wget -O test_tree.py ${SKLEARN_URL_ROOT}tree/tests/test_tree.py
+    ;;
+    *)
+        echo "Unknown algorithm: ${ALG_NAME}"
+        exit 1
+    ;;
+esac
