+++ .ci/conformance/run_tests.py
@@ -22,13 +22,10 @@ if __name__ == "__main__":
             Start of testing in """ + str(datetime.now())+ "<br>"
     report_file.write(textHTML)
 
-    os.system("bash ./download_tests.sh")
-
     for alg_name in algs:
         report_file.write("<br><h2>Testing %s</h2>" % alg_name)
-        #print("alg_name: ", alg_name)
-        #os.system("source ./tools/%s/download_test.sh &>> _log_downloads" % alg_name)
-        out, err = Popen("python tools/%(alg)s/patcher.py test_%(alg)s.py" % {"alg": alg_name}, shell=True, stdout=PIPE).communicate()
+        os.system("bash ./download_tests.sh --alg-name %s --scikit-version 0.21.3" % alg_name)
+        out, err = Popen("python ./patcher.py %(alg)s" % {"alg": alg_name}, shell=True, stdout=PIPE).communicate()
         report_file.write((str(out, 'utf-8')).replace("\n", "<br>"))
         os.system("pytest -s --disable-warnings test_%s.py > _log.txt" % alg_name)
         out, err = Popen("python tools/get_conformance_general.py", shell=True, stdout=PIPE).communicate()
@@ -37,9 +34,6 @@ if __name__ == "__main__":
             out, err = Popen("python tools/%s/get_conformance_local.py" % alg_name, shell=True, stdout=PIPE).communicate()
             report_file.write((str(out, 'utf-8')).replace("\n", "<br>"))
 
-        #os.system("rm test_%s.py" % alg_name)
-        #os.system("rm __n_calls.tmp")
-        #os.system("rm _log*")
 
     textHTML = """<br>
     Finishing testing in """+str(datetime.now())+"""<br></p>
