--- .ci/conformance/tools/adaboost/download_test.sh
@@ -1,4 +1,0 @@
-echo
-echo "Test adaboost download started."
-echo
-wget -O test_adaboost.py https://raw.githubusercontent.com/scikit-learn/scikit-learn/0.21.3/sklearn/ensemble/tests/test_weight_boosting.py
