--- .ci/conformance/tools/dbscan/download_test.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-echo
-echo "Test dbscan download started."
-echo
-wget -O test_dbscan.py https://raw.githubusercontent.com/scikit-learn/scikit-learn/master/sklearn/cluster/tests/test_dbscan.py
