--- .ci/conformance/tools/elastic_net/download_test.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-echo
-echo "Test elastic net regression download started."
-echo
-wget -O test_elastic_net.py https://raw.githubusercontent.com/scikit-learn/scikit-learn/0.21.3/sklearn/linear_model/tests/test_coordinate_descent.py
