--- .ci/conformance/tools/forest/download_test.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-echo
-echo "Test random_forest download started."
-echo
-wget -O test_forest.py https://raw.githubusercontent.com/scikit-learn/scikit-learn/0.21.3/sklearn/ensemble/tests/test_forest.py
