--- .ci/conformance/tools/kmeans/download_test.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-echo
-echo "Test kmeans download started."
-echo
-wget -O test_kmeans.py https://raw.githubusercontent.com/scikit-learn/scikit-learn/0.21.3/sklearn/cluster/tests/test_k_means.py
