--- .ci/conformance/tools/knn/download_test.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-echo
-echo "Test k-nearest neighbors download started."
-echo
-wget -O test_knn.py https://raw.githubusercontent.com/scikit-learn/scikit-learn/0.21.3/sklearn/neighbors/tests/test_neighbors.py
