--- .ci/conformance/tools/log_reg/download_test.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-echo
-echo "Test logistic regression download started."
-echo
-wget -O test_log_reg.py https://raw.githubusercontent.com/scikit-learn/scikit-learn/0.21.3/sklearn/linear_model/tests/test_logistic.py
