--- .ci/conformance/tools/pca/download_test.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-echo
-echo "Test pca download started."
-echo
-wget -O test_pca.py https://raw.githubusercontent.com/scikit-learn/scikit-learn/0.21.3/sklearn/decomposition/tests/test_pca.py
