--- .ci/conformance/tools/ridge_reg/download_test.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-echo
-echo "Test ridge regression download started."
-echo
-wget -O test_ridge_reg.py https://raw.githubusercontent.com/scikit-learn/scikit-learn/0.21.3/sklearn/linear_model/tests/test_ridge.py
