--- .ci/conformance/tools/svm/download_test.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-echo
-echo "Test svm download started."
-echo
-wget -O test_svm.py https://raw.githubusercontent.com/scikit-learn/scikit-learn/0.21.3/sklearn/svm/tests/test_svm.py
