--- .ci/conformance/tools/tree/download_test.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-echo
-echo "Test tree download started."
-echo
-wget -O test_tree.py https://raw.githubusercontent.com/scikit-learn/scikit-learn/0.21.3/sklearn/tree/tests/test_tree.py
